Move non Tally-specific code into a core data extension.
